feat(ext4): improvements in `Ext4File` and `Ext4Directory`, refactor
