+++ src/fs/ext4/dir.rs
@@ -3,19 +3,21 @@ use core::slice;
 use alloc::{string::String, vec::Vec};
 use bytemuck::{cast, from_bytes, try_cast, Pod, Zeroable};
 
+use crate::fs::ext4::file::Ext4File;
 use crate::{
     drivers::ahci::SATA_DRIVES,
     errors::{CanFail, IOError},
     fs::{
         ext4::{
             extent::{Ext4InodeRelBlkId, Ext4InodeRelBlkIdRange},
-            inode::{InodeFileMode, InodeSize},
-            Ext4File, Ext4Fs, ExtentTree, Inode,
+            inode::{InodeFileMode, InodeNumber, InodeSize},
+            Ext4Fs, ExtentTree, Inode,
         },
         IOResult, PartFS,
     },
 };
 
+/// Representation of a directory entry in the `ext4` filesystem.
 #[derive(Clone, Debug, Default, PartialEq, Eq, PartialOrd, Ord, Hash)]
 pub struct Ext4DirectoryEntry {
     drive_id: usize,
@@ -23,7 +25,11 @@ pub struct Ext4DirectoryEntry {
     inode: InodeNumber,
     rec_len: u16,
     name_len: u8,
+
+    /// File type associated to this entry (regular, directory, socket, ...)
     pub file_type: Option<Ext4DirectoryFileType>,
+
+    /// Name associated to this entry
     pub name: Ext4Filename,
 }
 
@@ -31,9 +37,13 @@ impl Ext4DirectoryEntry {
     /// Maximum ext4 directory entry size in bytes
     pub const MAX_ENTRY_SIZE: u16 = 263;
 
+    /// Consumes this `Ext4DirectoryEntry` into a [`Ext4Directory`].
+    ///
+    /// The file type associated with the entry must be [`Ext4DirectoryFileType::DIRECTORY`].
+    #[must_use]
     pub fn as_directory(&self) -> Option<Ext4Directory> {
-        if let Some(ftype) = self.file_type {
-            if ftype == Ext4DirectoryFileType::DIRECTORY {
+        if let Some(file_type) = self.file_type {
+            if file_type == Ext4DirectoryFileType::DIRECTORY {
                 return Ext4Directory::from_inode_id(self.drive_id, self.part_id, self.inode).ok();
             }
         }
@@ -41,73 +51,18 @@ impl Ext4DirectoryEntry {
         None
     }
 
+    /// Consumes this `Ext4DirectoryEntry` into a [`Ext4File`].
+    ///
+    /// The file type associated with the entry must be [`Ext4DirectoryFileType::REGULAR`].
+    #[must_use]
     pub fn as_file(&self) -> Option<Ext4File> {
-        if let Some(ftype) = self.file_type {
-            if ftype == Ext4DirectoryFileType::REGULAR {
-                return Ext4File::from_inode_id(
-                    self.drive_id,
-                    self.part_id,
-                    u32::from(self.inode) as usize,
-                )
-                .ok();
+        if let Some(file_type) = self.file_type {
+            if file_type == Ext4DirectoryFileType::REGULAR {
+                return Ext4File::from_inode_id(self.drive_id, self.part_id, self.inode).ok();
             }
         }
 
         None
-    }
-}
-
-/// A number representing an inode.
-#[derive(Clone, Copy, Debug, Default, Eq, PartialEq, PartialOrd, Ord, Hash, Pod, Zeroable)]
-#[repr(C)]
-pub struct InodeNumber(u32);
-
-impl InodeNumber {
-    /// Inode 0 represents an unused directory entry
-    pub const UNUSED_DIR_ENTRY: Self = Self(0);
-
-    /// Inode 2 is reserved for the root directory of the file system.
-    pub const ROOT_DIR: Self = Self(0x2);
-
-    /// Inode 3 is reserved for the user quota
-    pub const USER_QUOTA: Self = Self(0x3);
-
-    /// Inode 4 is reserved for the group quota file
-    pub const GROUP_QUOTA: Self = Self(0x4);
-
-    /// Inode 5 is unused, but may have been intended for stage 2 bootloaders
-    pub const BOOTLOADER: Self = Self(0x5);
-
-    /// Inode 6 is unused, but may have been intended for the never implemented undeletion
-    pub const UNDELETE: Self = Self(0x6);
-
-    /// Inode 7 is the reserved group descriptors inode
-    pub const RESIZE: Self = Self(0x7);
-
-    /// Inode 8 is the ext4 journal
-    pub const JOURNAL: Self = Self(0x8);
-
-    /// Inode 9 is the exclude inode, for snapshots
-    pub const EXCLUDE: Self = Self(0x9);
-
-    /// Inode 10 is used for ext4 metadata replication in some non-upstream patches
-    pub const REPLICA: Self = Self(0xA);
-}
-
-impl From<InodeNumber> for u32 {
-    fn from(value: InodeNumber) -> Self {
-        value.0
-    }
-}
-
-impl From<InodeNumber> for usize {
-    fn from(value: InodeNumber) -> usize {
-        value.0.try_into().expect("invalid inode number")
-    }
-}
-impl From<usize> for InodeNumber {
-    fn from(value: usize) -> Self {
-        InodeNumber(value.try_into().expect("invalid inode number (not 32-bit)"))
     }
 }
 
@@ -134,10 +89,14 @@ impl Ext4DirectoryFileType {
     pub const SYMLINK: Self = Self(0x7);
 }
 
+/// File name associated to a [`Ext4DirectoryEntry`].
+///
+/// UTF-8 encoded string.
 #[derive(Clone, Debug, Default, Eq, PartialEq, PartialOrd, Ord, Hash)]
 pub struct Ext4Filename(pub Vec<u8>);
 
 impl Ext4Filename {
+    /// Returns an [`Iterator`] over the characters of the file name.
     pub fn chars(&self) -> impl Iterator<Item = char> {
         self.0
             .clone()
@@ -149,7 +108,7 @@ impl Ext4Filename {
 
 impl From<Ext4Filename> for String {
     fn from(value: Ext4Filename) -> Self {
-        String::from_iter(value.chars())
+        value.chars().collect::<String>()
     }
 }
 
@@ -159,6 +118,8 @@ impl From<&str> for Ext4Filename {
     }
 }
 
+/// Representation of a directory in the `ext4` filesystem.
+#[derive(Clone, Debug, Ord, PartialOrd, Eq, PartialEq, Hash)]
 pub struct Ext4Directory {
     drive_id: usize,
     part_id: usize,
@@ -175,7 +136,7 @@ impl Iterator for Ext4Directory {
         let mut raw_entry = [0u8; Ext4DirectoryEntry::MAX_ENTRY_SIZE as usize];
         let count_to_read = u64::min(
             cast(self.inode.size() - self.internal_cursor.try_into().ok()?),
-            Ext4DirectoryEntry::MAX_ENTRY_SIZE as u64,
+            u64::from(Ext4DirectoryEntry::MAX_ENTRY_SIZE),
         )
         .try_into()
         .ok()?;
@@ -219,10 +180,19 @@ impl Iterator for Ext4Directory {
 }
 
 impl Ext4Directory {
+    /// Search this directory for a given [`Ext4Filename`].
+    ///
+    /// Returns the corresponding entry if available.
     pub fn search(&mut self, name: Ext4Filename) -> Option<Ext4DirectoryEntry> {
         self.find(|entry| entry.name == name)
     }
 
+    /// Loads a `Ext4Directory` from disk, from its [`InodeNumber`].
+    ///
+    /// # Errors
+    ///
+    /// May return any variant of [`IOError`] in case of a failure while attempting to read from disk, or because of
+    /// an invalid disk descriptor.
     pub fn from_inode_id(drive_id: usize, part_id: usize, inode_id: InodeNumber) -> IOResult<Self> {
         let drive = SATA_DRIVES
             .get()
@@ -238,13 +208,20 @@ impl Ext4Directory {
 
         if let PartFS::Ext4(fs) = fs {
             drop(drive);
-            let inode = fs.__read_inode(u32::from(inode_id) as u64)?;
+            let inode = fs.__read_inode(inode_id)?;
 
             return Self::from_inode(drive_id, part_id, inode, inode_id);
         }
 
         Err(IOError::Unknown)
     }
+
+    /// Loads a `Ext4Directory` from disk, from its [`InodeNumber`] and the corresponding [`Inode`] structure.
+    ///
+    /// # Errors
+    ///
+    /// May return any variant of [`IOError`] in case of a failure while attempting to read from disk, or because of
+    /// an invalid disk descriptor.
     pub(crate) fn from_inode(
         drive_id: usize,
         part_id: usize,
@@ -316,8 +293,7 @@ impl Ext4Directory {
 
         if let Some(ext_tree) = &self.extent_tree {
             let mut useful_extents = ext_tree.extents.iter().filter(|ext| {
-                (cast(ext.ee_block)..ext.ee_block + ext.ee_len)
-                    .contains(&(blk_offset_from_file_start))
+                (cast(ext.block)..ext.block + ext.len).contains(&(blk_offset_from_file_start))
             });
             let mut curr_extent = useful_extents.next().unwrap();
 
@@ -325,51 +301,62 @@ impl Ext4Directory {
                 blk_offset_from_file_start,
                 Ext4InodeRelBlkId::min(cast(0_u64), last_blk - 1),
             ) {
-                if (curr_extent.ee_block + curr_extent.ee_len) < i {
+                if (curr_extent.block + curr_extent.len) < i {
                     curr_extent = useful_extents.next().unwrap();
                 }
                 fs.__read_blk(
                     try_cast(curr_extent.start_blk() + i).map_err(|_| IOError::Unknown)?,
                     slice::from_raw_parts_mut(
-                        buf.as_mut_ptr().offset(
-                            (try_cast::<Ext4InodeRelBlkId, u64>(i).map_err(|_| IOError::Unknown)?
-                                * fs.superblock.blk_size()) as isize,
+                        buf.as_mut_ptr().add(
+                            (try_cast::<Ext4InodeRelBlkId, u64>(i)
+                                .map_err(|_| IOError::Unknown)?
+                                * fs.superblock.blk_size())
+                            .try_into()
+                            .expect("invalid inode number"),
                         ),
-                        fs.superblock.blk_size() as usize,
+                        fs.superblock
+                            .blk_size()
+                            .try_into()
+                            .expect("invalid fs block size"),
                     ),
                 )?;
             }
 
-            if (curr_extent.ee_block + curr_extent.ee_len) < last_blk {
+            if (curr_extent.block + curr_extent.len) < last_blk {
                 curr_extent = useful_extents.next().unwrap();
             }
 
-            let mut temp_buf = alloc::vec![0u8; fs.superblock.blk_size() as usize];
+            let mut temp_buf = alloc::vec![0u8; fs.superblock
+                .blk_size().try_into().expect("invalid fs block size")];
             fs.__read_blk(
                 try_cast(curr_extent.start_blk() + last_blk).map_err(|_| IOError::Unknown)?,
                 &mut temp_buf,
             )?;
 
-            if last_blk != cast::<u64, Ext4InodeRelBlkId>(0) {
+            if last_blk == cast::<u64, Ext4InodeRelBlkId>(0) {
                 slice::from_raw_parts_mut(
-                    buf.as_mut_ptr().offset(
+                    buf.as_mut_ptr().add(
                         (try_cast::<Ext4InodeRelBlkId, u64>(last_blk)
                             .map_err(|_| IOError::Unknown)?
-                            * fs.superblock.blk_size()) as isize,
-                    ),
-                    last_blk_count,
-                )
-                .copy_from_slice(&temp_buf[..last_blk_count]);
-            } else {
-                slice::from_raw_parts_mut(
-                    buf.as_mut_ptr().offset(
-                        (try_cast::<Ext4InodeRelBlkId, u64>(last_blk)
-                            .map_err(|_| IOError::Unknown)?
-                            * fs.superblock.blk_size()) as isize,
+                            * fs.superblock.blk_size())
+                        .try_into()
+                        .expect("invalid inode number"),
                     ),
                     last_blk_count - blk_offset_from_first_blk,
                 )
                 .copy_from_slice(&temp_buf[blk_offset_from_first_blk..last_blk_count]);
+            } else {
+                slice::from_raw_parts_mut(
+                    buf.as_mut_ptr().add(
+                        (try_cast::<Ext4InodeRelBlkId, u64>(last_blk)
+                            .map_err(|_| IOError::Unknown)?
+                            * fs.superblock.blk_size())
+                        .try_into()
+                        .expect("invalid inode number"),
+                    ),
+                    last_blk_count,
+                )
+                .copy_from_slice(&temp_buf[..last_blk_count]);
             }
         }
